requirements.txt
uvicorn[standard]~=0.23.2
cloudscraper~=1.2.71
fastapi~=0.103.1
beautifulsoup4~=4.12.2
celery~=5.3.4
python-dotenv~=1.0.0